Move http to http_util.
