+++ afscgap/test/test_inference.py
@@ -1,3 +1,12 @@
+"""
+Tests for zero catch inference utilities.
+
+(c) 2023 Regents of University of California / The Eric and Wendy Schmidt Center
+for Data Science and the Environment at UC Berkeley.
+
+This file is part of afscgap released under the BSD 3-Clause License. See
+LICENSE.md.
+"""
 import unittest
 import unittest.mock
